+++ evm/evm-processor/src/mapping/schema.ts
@@ -1,5 +1,17 @@
 import {FieldSelection} from '../interfaces/data'
-import {array, BYTES, NAT, object, option, QTY, SMALL_QTY, STRING, taggedUnion, withSentinel} from '@subsquid/util-internal-validation'
+import {
+    array,
+    BYTES,
+    NAT,
+    object,
+    option,
+    QTY,
+    SMALL_QTY,
+    STRING,
+    taggedUnion,
+    withDefault,
+    withSentinel
+} from '@subsquid/util-internal-validation'
 
 
 export function getBlockHeaderProps(fields: FieldSelection['block'], forArchive: boolean) {
@@ -20,7 +32,7 @@ export function getBlockHeaderProps(fields: FieldSelection['block'], forArchive:
             difficulty: withSentinel('BlockHeader.difficulty', -1n, QTY),
             totalDifficulty: withSentinel('BlockHeader.totalDifficulty', -1n, QTY),
             extraData: withSentinel('BlockHeader.extraData', '0x', BYTES),
-            size: withSentinel('BlockHeader.size', -1, SMALL_QTY),
+            size: withSentinel('BlockHeader.size', -1, natural),
             gasLimit: withSentinel('BlockHeader.gasLimit', -1n, QTY),
             gasUsed: withSentinel('BlockHeader.gasUsed', -1n, QTY),
             baseFeePerGas: withSentinel('BlockHeader.baseFeePerGas', -1n, QTY),
@@ -137,7 +149,7 @@ export function getTraceFrameValidator(fields: FieldSelection['trace'], forArchi
         value: option(QTY),
         gas: QTY,
         input: BYTES,
-        sighash: BYTES
+        sighash: withDefault('0x', BYTES)
     })
 
     let traceCallResult = project({