+++ graphql/graphql-server/src/test/util/server.ts
@@ -1,6 +1,6 @@
-import {ListeningServer} from "@subsquid/util-internal-http-server"
-import {Client} from "gql-test-client"
-import {Server} from "../../server"
+import {ListeningServer} from '@subsquid/util-internal-http-server'
+import {Client} from 'gql-test-client'
+import {Server} from '../../server'
 
 
 export function useServer(project: string): Client {
@@ -8,7 +8,10 @@ export function useServer(project: string): Client {
     let info: ListeningServer | undefined
 
     before(async () => {
-        info = await new Server({dir: project}).start()
+        info = await new Server({
+            dir: project,
+            // log: createLogger('sqd:graphql-server')
+        }).start()
         client.endpoint = `http://localhost:${info.port}/graphql`
     })