+++ graphql/openreader/src/context.ts
@@ -1,6 +1,7 @@
-import {Dialect} from './dialect'
-import {Query} from './sql/query'
-import {Limit} from './util/limit'
+import type {Logger} from '@subsquid/logger'
+import type {Dialect} from './dialect'
+import type {Query} from './sql/query'
+import type {Limit} from './util/limit'
 
 
 export interface Context {
@@ -9,9 +10,11 @@ export interface Context {
 
 
 export interface OpenreaderContext {
+    id: number
     dialect: Dialect
     executeQuery<T>(query: Query<T>): Promise<T>
     subscription<T>(query: Query<T>): AsyncIterable<T>
     responseSizeLimit?: Limit
     subscriptionResponseSizeLimit?: Limit
+    log?: Logger
 }