+++ substrate/substrate-processor/src/processor.ts
@@ -76,9 +76,9 @@ export interface RpcDataIngestionSettings {
 }
 
 
-export interface ArchiveSettings {
+export interface GatewaySettings {
     /**
-     * Subsquid archive URL
+     * Subsquid Network Gateway url
      */
     url: string
     /**
@@ -86,6 +86,12 @@ export interface ArchiveSettings {
      */
     requestTimeout?: number
 }
+
+
+/**
+ * @deprecated
+ */
+export type ArchiveSettings = GatewaySettings
 
 
 /**
@@ -147,7 +153,7 @@ export class SubstrateBatchProcessor<F extends FieldSelection = {}> {
     private requests: RangeRequest<DataRequest>[] = []
     private fields?: FieldSelection
     private blockRange?: Range
-    private archive?: ArchiveSettings
+    private archive?: GatewaySettings
     private rpcEndpoint?: RpcEndpointSettings
     private rpcIngestSettings?: RpcDataIngestionSettings
     private typesBundle?: OldTypesBundle | OldSpecsBundle
@@ -155,15 +161,22 @@ export class SubstrateBatchProcessor<F extends FieldSelection = {}> {
     private running = false
 
     /**
-     * Set Subsquid Archive endpoint.
+     * @deprecated Use {@link .setGateway()}
+     */
+    setArchive(url: string | GatewaySettings): this {
+        return this.setGateway(url)
+    }
+
+    /**
+     * Set Subsquid Network Gateway endpoint (ex Archive).
      *
-     * Subsquid Archive allows to get data from finalized blocks up to
+     * Subsquid Network allows to get data from finalized blocks up to
      * infinite times faster and more efficient than via regular RPC.
      *
      * @example
-     * processor.setArchive('https://v2.archive.subsquid.io/network/kusama')
+     * processor.setGateway('https://v2.archive.subsquid.io/network/kusama')
      */
-    setArchive(url: string | ArchiveSettings): this {
+    setGateway(url: string | GatewaySettings): this {
         this.assertNotRunning()
         if (typeof url == 'string') {
             this.archive = {url}
@@ -205,13 +218,13 @@ export class SubstrateBatchProcessor<F extends FieldSelection = {}> {
      *     chain: 'https://kusama-rpc.polkadot.io'
      * })
      *
-     * @deprecated Use separate {@link .setArchive()} and {@link .setRpcEndpoint()} methods
+     * @deprecated Use separate {@link .setGateway()} and {@link .setRpcEndpoint()} methods
      * to specify data sources.
      */
     setDataSource(src: DataSource): this {
         this.assertNotRunning()
         if (src.archive) {
-            this.setArchive(src.archive)
+            this.setGateway(src.archive)
         } else {
             this.archive = undefined
         }