+++ graphql/graphql-server/CHANGELOG.md
@@ -1,6 +1,26 @@
 # Change Log - @subsquid/graphql-server
 
-This log was last generated on Fri, 01 Dec 2023 16:55:51 GMT and should not be manually modified.
+This log was last generated on Sun, 24 Dec 2023 20:05:50 GMT and should not be manually modified.
+
+## 4.5.0
+Sun, 24 Dec 2023 20:05:50 GMT
+
+### Minor changes
+
+- support loading sever extensions via `ts-node`
+- implement query log
+- enhance graphql error log
+
+### Patches
+
+- fix class-validator@14 regression: an unknown value was passed to the validate function
+
+## 4.4.0
+Thu, 14 Dec 2023 11:16:15 GMT
+
+### Minor changes
+
+- introduce `DB_URL`, `DB_SSL_CA(_FILE)`, `DB_SSL_CERT(_FILE)`, `DB_SSL_KEY(_FILE)` env variables
 
 ## 4.3.2
 Fri, 01 Dec 2023 16:55:51 GMT