+++ typeorm/typeorm-migration/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @subsquid/typeorm-migration
 
-This log was last generated on Fri, 01 Dec 2023 16:55:51 GMT and should not be manually modified.
+This log was last generated on Sun, 24 Dec 2023 20:05:50 GMT and should not be manually modified.
+
+## 1.3.0
+Sun, 24 Dec 2023 20:05:50 GMT
+
+### Minor changes
+
+- support `ts-node`
+
+## 1.2.4
+Thu, 14 Dec 2023 11:16:15 GMT
+
+_Version update only_
 
 ## 1.2.3
 Fri, 01 Dec 2023 16:55:51 GMT