+++ util/rpc-client/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @subsquid/rpc-client
 
-This log was last generated on Fri, 01 Dec 2023 16:55:51 GMT and should not be manually modified.
+This log was last generated on Sun, 24 Dec 2023 20:05:50 GMT and should not be manually modified.
+
+## 4.6.0
+Sun, 24 Dec 2023 20:05:50 GMT
+
+### Minor changes
+
+- annotate rpc errors with endpoint url and response body
 
 ## 4.5.0
 Fri, 01 Dec 2023 16:55:51 GMT